move SpaceCard to ui folder
