+++ apps/client/app/(navbar)/user/[id]/Spaces.tsx
@@ -1,5 +1,5 @@
 import { fetchLatestSpaces } from "../../../../src/server/helpers/fetchLatestSpaces";
-import SpaceCard from "../../explore/SpaceCard";
+import SpaceCard from "../../../../src/ui/SpaceCard";
 
 interface Props {
   owner: string;
