+++ apps/client/src/ui/SpaceCard.tsx
@@ -1,9 +1,9 @@
 import { ERC721Metadata } from "contracts";
 import Link from "next/link";
 
-import Card from "../../../src/ui/Card";
-import { toHex } from "../../../src/utils/toHex";
-import PlayerCount from "./PlayerCount";
+import PlayerCount from "../../app/(navbar)/explore/PlayerCount";
+import { toHex } from "../utils/toHex";
+import Card from "./Card";
 
 interface Props {
   id: number;
@@ -11,6 +11,11 @@ interface Props {
   sizes?: string;
 }
 
+/**
+ * Wrapper around {@link Card} that links to the space page, and shows the player count.
+ * @param id Space ID
+ * @param metadata Space metadata
+ */
 export default function SpaceCard({ id, metadata, sizes }: Props) {
   return (
     <Link href={`/space/${toHex(id)}`} className="rounded-xl">
